% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{beginStyleFmt}
\alias{beginStyleFmt}
\alias{endStyleFmt}
\title{Style format options for longer text output}
\usage{
beginStyleFmt(textstyle, type = "span")

endStyleFmt(textstyle, type = "span")
}
\arguments{
\item{textstyle}{}

\item{type}{}

\item{textstyle}{}

\item{type}{}
}
\description{
Style format options for longer text output
}
