% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{beginVerbatim}
\alias{beginVerbatim}
\alias{endVerbatim}
\title{Specific to verbatim environment 
to be able to show you 'asis' chunks with background color.}
\usage{
beginVerbatim()

endVerbatim()
}
\description{
Specific to verbatim environment 
to be able to show you 'asis' chunks with background color.
}
