% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{colFmt}
\alias{colFmt}
\title{Simple color Format functions}
\usage{
colFmt(x, colorname, type = "span")
}
\arguments{
\item{x}{}

\item{colorname}{}

\item{type}{}
}
\description{
Simple color Format functions
}
