% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{styleFmt}
\alias{styleFmt}
\title{Style format function, not specific for textcolor}
\usage{
styleFmt(x, textstyle, type = "span")
}
\arguments{
\item{x}{}

\item{textstyle}{}

\item{type}{}
}
\description{
Style format function, not specific for textcolor
}
