% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{colFmt}
\alias{colFmt}
\title{Simple color format functions}
\usage{
colFmt(x, colorname, type = "span")
}
\arguments{
\item{x}{texte to be formatted}

\item{colorname}{colorname as known by latex or html (red, blue, ...)}

\item{type}{html output format. Default to span.}
}
\description{
Simple color format functions
}
